# docs/swagger.yaml
definitions:
  domain.CreateProductRequest:
    properties:
      description:
        type: string
      name:
        type: string
      price:
        type: integer
      qty:
        type: integer
    type: object
  domain.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  domain.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      role_id:
        type: integer
    type: object
  domain.UpdateProductRequest:
    properties:
      description:
        type: string
      name:
        type: string
      price:
        type: integer
      qty:
        type: integer
    type: object
  util.Base:
    properties:
      data:
        type: object
      message:
        type: string
      status:
        type: string
      status_code:
        type: integer
      timestamp:
        type: string
    type: object
info:
  contact: {}
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login User
      parameters:
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Base'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Base'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.Base'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.Base'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Base'
      summary: Login User
      tags:
      - Authentication
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register User
      parameters:
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Base'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Base'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.Base'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.Base'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Base'
      summary: Register User
      tags:
      - Authentication
  /products:
    get:
      consumes:
      - application/json
      description: GetProducts User
      parameters:
      - description: Bearer (token)
        in: header
        name: Authorization
        required: true
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: sort by
        in: query
        name: sort_by
        type: string
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Base'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Base'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.Base'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.Base'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Base'
      summary: GetProducts User
      tags:
      - Product
    post:
      consumes:
      - application/json
      description: CreateProduct User
      parameters:
      - description: Bearer (token)
        in: header
        name: Authorization
        required: true
        type: string
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.CreateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Base'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Base'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.Base'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.Base'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Base'
      summary: CreateProduct User
      tags:
      - Product
  /products/{product_id}:
    delete:
      consumes:
      - application/json
      description: ForceDeleteProduct User
      parameters:
      - description: Bearer (token)
        in: header
        name: Authorization
        required: true
        type: string
      - description: product id
        in: path
        name: product_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Base'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Base'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.Base'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.Base'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Base'
      summary: ForceDeleteProduct User
      tags:
      - Product
    get:
      consumes:
      - application/json
      description: GetProduct User
      parameters:
      - description: Bearer (token)
        in: header
        name: Authorization
        required: true
        type: string
      - description: product id
        in: path
        name: product_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Base'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Base'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.Base'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.Base'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Base'
      summary: GetProduct User
      tags:
      - Product
    put:
      consumes:
      - application/json
      description: UpdateProduct User
      parameters:
      - description: Bearer (token)
        in: header
        name: Authorization
        required: true
        type: string
      - description: product id
        in: path
        name: product_id
        required: true
        type: string
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Base'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Base'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.Base'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.Base'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Base'
      summary: UpdateProduct User
      tags:
      - Product
swagger: "2.0"